wallet: implement `ListAddresses`
The new `ListAddresses` method is a significant improvement over the
previous implementation found in `lnd`. The old approach had a time
complexity of O(U * A) (UTXOs * Addresses), as it would iterate through
all UTXOs for each address to calculate its balance.

The new implementation is much more efficient with a time complexity of
O(U + A). It achieves this by first iterating through all UTXOs once to
create a balance map, and then iterating through the addresses to look
up their balances in the map. This is all done within a single,
read-only database transaction.