wallet: add `TxWriter` interface and implementation
This commit introduces a new `TxWriter` interface that provides methods
for updating wallet transactions. The first method on this interface is
`LabelTx`, which allows a caller to add or overwrite a label for a given
transaction.

The `Wallet` type now implements this interface. This change also
includes unit tests to verify the functionality of labeling a
transaction, including success and error cases.